+++ artipie-core/src/main/java/com/artipie/misc/Property.java
@@ -5,6 +5,7 @@
 package com.artipie.misc;
 
 import com.artipie.ArtipieException;
+
 import java.util.Optional;
 
 /**
