+++ artipie-core/src/test/resources/artipie.properties
@@ -1,0 +1,5 @@
+artipie.version=${project.version}
+artipie.config.cache.timeout=120000
+artipie.cached.auth.timeout=300000
+artipie.storage.file.cache.timeout=180000
+artipie.credentials.file.cache.timeout=180000
