+++ artipie-main/src/test/java/com/artipie/test/TestArtipieCaches.java
@@ -7,7 +7,7 @@ package com.artipie.test;
 import com.artipie.asto.misc.Cleanable;
 import com.artipie.settings.cache.ArtipieCaches;
 import com.artipie.settings.cache.FiltersCache;
-import com.artipie.settings.cache.StoragesCache;
+
 import java.util.concurrent.atomic.AtomicLong;
 import org.apache.commons.lang3.NotImplementedException;
